+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+# [2.0.0-beta.4](https://github.com/mob-sakai/git-snapshot/compare/v2.0.0-beta.3@beta...v2.0.0-beta.4@beta) (2019-10-16)
+
+
+### Features
+
+* test build ([0a1ddb9](https://github.com/mob-sakai/git-snapshot/commit/0a1ddb9b5f9ddf3a94c218c6840d5ef45438cccf))
+
+# [2.0.0-beta.3](https://github.com/mob-sakai/git-snapshot/compare/v2.0.0-beta.2@beta...v2.0.0-beta.3@beta) (2019-10-16)
+
+
+### Features
+
+* test build ([02f0588](https://github.com/mob-sakai/git-snapshot/commit/02f05883fe95fadaee5c8e958fd7dd311f2fdebb))
+
+# [2.0.0-beta.2](https://github.com/mob-sakai/git-snapshot/compare/v2.0.0-beta.1@beta...v2.0.0-beta.2@beta) (2019-10-16)
+
+
+### Bug Fixes
+
+* Exception object ([d0d2b1a](https://github.com/mob-sakai/git-snapshot/commit/d0d2b1a7c7287ee6d892a17bc36c9c59101e2023))
+* Prefix dir ([7600e99](https://github.com/mob-sakai/git-snapshot/commit/7600e996536860c9d50c8fb2109cf199b423792f))
+
+
+### Features
+
+* Check --prefix and --cwd options ([8566a9f](https://github.com/mob-sakai/git-snapshot/commit/8566a9fcf984908211b21e1145aa28ac3071c5b9))
+* Support dry-run mode ([53c5ec0](https://github.com/mob-sakai/git-snapshot/commit/53c5ec0e85051ee31df6d16a144df62899637849))
+
+# [2.0.0-beta.1](https://github.com/mob-sakai/git-snapshot/compare/v1.1.2...v2.0.0-beta.1@beta) (2019-10-16)
+
+
+### Features
+
+* Add cli ([1ed7ada](https://github.com/mob-sakai/git-snapshot/commit/1ed7ada7ef1883a998e3ecda40ffdd98f39b3c0d))
+* Add git function ([96d95f9](https://github.com/mob-sakai/git-snapshot/commit/96d95f997c4c670c967b4bb751737c68873350eb))
+* Add more options (allow-empty, author, force, tag, remote, dry-run, cwd) ([29e6208](https://github.com/mob-sakai/git-snapshot/commit/29e62084aa9ad10baf8a59343eea4c8eed7fafc2))
+* Rewrite in javascript ([75933a0](https://github.com/mob-sakai/git-snapshot/commit/75933a0999f556b2e2a2ed340fb9d516271eca3b))
+
+
+### BREAKING CHANGES
+
+* Shell script has been removed.
+Available as a package.
+
 ## [1.1.2](https://github.com/mob-sakai/git-snapshot/compare/1.1.1...1.1.2) (2019-07-16)